+++ src/FPrimitive/Access.fs
@@ -24,32 +24,43 @@ module Access =
     let funcResult f = { Capability = f; Revokable = None }
 
     /// Creates an access controlled function from a function, directly using the returned `Option`.
-    let funcOption error f = { Capability = f >> Result.ofOption error; Revokable = None }
+    let funcOption f error = { Capability = f >> Result.ofOption [error]; Revokable = None }
 
+    /// Creates an access-controlled function from a custom function.
     [<CompiledName("Function")>]
     let Function2 (f : Func<_, _>) = 
         if f = null then nullArg "f"
         func f.Invoke
 
+    /// Creates an access-controlled function from a custom function.
     [<CompiledName("Function")>]
     let Function1 (f : Func<_>) =
         if f = null then nullArg "f"
         func f.Invoke
 
+    /// Creates an access controlled function from a custom function.
     [<CompiledName("Function")>]
     let Action (f : Action<_>) =
         if f = null then nullArg "f"
         func f.Invoke
 
+    /// Adds a custom requirement that the access-controlled function.
     let filter predicate error acc = 
-        { acc with Capability = acc.Capability >> Result.filter predicate [ "access failure:" + error ] }
+        { acc with 
+            Capability = fun x -> 
+              if predicate x then acc.Capability x 
+              else Error [ "access failure:" + error ] }
     
+    /// Evaluate the access-controlled function.
     let eval x { Capability = f } = f x
  
+    /// Evaluate the access-controlled function with unit `()`.
     let evalUnit acc = acc.Capability ()
 
+    /// Revokes the access-controlled function.
     let revoke acc = Option.iter (fun f -> f ()) acc.Revokable
 
+    /// Creates an access-controlled function that only retrieves files from certain directories.
     let onlyFilesFromDirs (expected : DirectoryInfo seq) =
         let expected = List.ofSeq expected |> List.map (fun (x : DirectoryInfo) -> Path.GetFullPath x.FullName)
         { Revokable = None
@@ -60,13 +71,16 @@ module Access =
             else let dirs = String.Join (", ", Array.ofList expected)
                  Error [ sprintf "access failure: only files from directories: '%s' are allowed, but came from '%s'" dirs actualPath ] }
 
-    let OnlyFilesFromDirectories ([<ParamArray] expected : DirectoryInfo []) =
+    /// Creates an access-controlled function that only retrieves files from certain directories.
+    let OnlyFilesFromDirectories ([<ParamArray>] expected : DirectoryInfo []) =
       onlyFilesFromDirs (Seq.ofArray expected)
 
+    /// Creates an access-controlled function that only retrieves files from a certain directory.
     [<CompiledName("OnlyFilesFromDirectory")>]
     let onlyFilesFrom expected = onlyFilesFromDirs [expected]
 
-    let extensions exts acc =
+    /// Adds a requirement to the access-controlled function to only allow files with certain extensions.
+    let fileExtensions exts acc =
         { acc with
             Capability = fun (actual : FileInfo) ->
                 if List.contains actual.Extension exts
@@ -74,8 +88,10 @@ module Access =
                 else let exts = String.Join (", ", Array.ofList exts)
                      Error [ sprintf "access failure: only files with extension: '%s' are allowed" exts ] }
 
-    let extension ext acc = extensions [ext] acc
+    /// Adds a requirement to the access-controlled function to only allow files with a certain extension.
+    let fileExtension ext acc = fileExtensions [ext] acc
 
+    /// Creates an access-controlled function to only allow `Uri` values with a certain base `Uri`.
     [<CompiledName("OnlyUriFromBase")>]
     let onlyUriFromBase baseUri =
         { Revokable = None
@@ -84,6 +100,7 @@ module Access =
             then Ok actual
             else Error [ sprintf "access failure: only URI's with the base of '%s' are allowed, but got: '%s'" baseUri.OriginalString actual.OriginalString ] }
 
+    /// Let the access-controlled function only evaluate once.
     let once acc =
         let func = ref (Func<_, _> acc.Capability)
         { acc with 
@@ -92,6 +109,7 @@ module Access =
                 if func = null then Error [ "access failure: function is already evaluated" ]
                 else func.Invoke x }
 
+    /// Let the access-controlled function only evaluate a certain amount of times.
     let times count acc =
         if count < 0 then invalidArg "count" "amount of times to access resource must be greater than zero"
         let count = ref count
@@ -105,8 +123,10 @@ module Access =
                   else func.Invoke x
                 else (!func).Invoke x }
 
+    /// Let the access-controlled function only evaluate two times.
     let twice acc = times 2 acc
 
+    /// Make the access-controlled function revokable.
     let revokable acc =
         let allow = ref (true :> obj)
         let capability x =
@@ -115,6 +135,7 @@ module Access =
         { Capability = capability
           Revokable = Some (fun () -> Interlocked.Exchange (allow, false) |> ignore) }
 
+    /// Automatically revoke the access-controlled after a specified delay.
     let revokedAfter delay acc =
         let acc = revokable acc
         let timer = new Timers.Timer ()
@@ -127,11 +148,13 @@ module Access =
         timer.Start ()
         acc
 
+    /// Automatically revoke the access-controlled function when the specified observable emits a value.
     let revokedWhen observable acc =
       let acc = revokable acc
       Observable.add (fun _ -> revoke acc) observable
       acc
 
+    /// Make the access-controlled function only available during certain dates.
     let duringDates min max acc =
         { acc with
             Capability = fun x ->
@@ -140,130 +163,162 @@ module Access =
                 then acc.Capability x
                 else Error [ sprintf "access failure: can only access resource between '%A' and '%A', but now it's '%A'" min max now ] }
 
+    /// Make the access-controlled function only available during certain hours.
     let duringHours min max acc =
         if min < 0 || min > 23 then invalidArg "min" "minimum hour should be in the range of 0-23"
         if max < 0 || max > 23 then invalidArg "max" "maximum hour should be in the range of 0-23"
         if min > max then invalidArg "min" "minimum hour cannot be greater than maximum hour" 
         { acc with
             Capability = fun x ->
-                let now = DateTimeOffset.Now.Hour
-                if min < now && now > max
+                let now = DateTimeOffset.Now
+                if min < now.Hour && now.Hour < max
                 then acc.Capability x
-                else Error [ sprintf "access failure: can only access resource between '%A' and '%A', but now it's '%A'" min max now ] }
+                else Error [ sprintf "access failure: can only access resource between '%i:00' and '%i:00', but now it's '%i:%i'" min max now.Hour now.Minute ] }
     
-
-    let matching spec acc =
+    /// Adds input-validation to the access-controlled function.
+    let satisfy spec acc =
       { acc with
           Capability = fun x -> 
             match Spec.validate x spec with
             | Ok x -> acc.Capability x
             | Error errs -> Error errs }
 
+/// Computation expression to control the access of functions.
 type AccessBuilder<'a, 'b> () =
     member __.Yield (_) = 
         { Revokable = None
           Capability = fun _ -> Error [ "access failure: no resource was configured" ] }
+    /// Creates an access controlled function from a custom function.
     [<CustomOperation("func")>]
     member __.Func (_, f) = Access.func f
+    /// Creates an access-controlled function from a function, directly using the returned `Result`.
     [<CustomOperation("funcResult")>]
     member __.FuncResult (_, f) = Access.funcResult f
+    /// Creates an access controlled function from a function, directly using the returned `Option`.
     [<CustomOperation("funcOption")>]
-    member __.FuncOption (_, f) = Access.funcOption f
+    member __.FuncOption (_, f, error) = Access.funcOption f error
+    /// Creates an access-controlled function that only retrieves files from a certain directory.
     [<CustomOperation("onlyFilesFrom")>]
     member __.OnlyFilesFrom (_, dir) = Access.onlyFilesFrom dir
+    /// Creates an access-controlled function that only retrieves files from certain directories.
     [<CustomOperation("onlyFilesFromDirs")>]
     member __.OnlyFilesFromDirs (_, dirs) = Access.onlyFilesFromDirs dirs
-    [<CustomOperation("extension")>]
-    member __.Extension (state, ext) = Access.extension ext state
-    [<CustomOperation("extensions")>]
-    member __.Extensions (state, exts) = Access.extensions exts state
+    /// Adds a requirement to the access-controlled function to only allow files with a certain extension.
+    [<CustomOperation("fileExtension")>]
+    member __.Extension (state, ext) = Access.fileExtension ext state
+    /// Adds a requirement to the access-controlled function to only allow files with certain extensions.
+    [<CustomOperation("filesExtensions")>]
+    member __.Extensions (state, exts) = Access.fileExtensions exts state
+    /// Creates an access-controlled function to only allow `Uri` values with a certain base `Uri`.
     [<CustomOperation("onlyUriFromBase")>]
     member __.OnlyUriFrombase (_, baseUri) = Access.onlyUriFromBase baseUri
     [<CustomOperation("filter")>]
     member __.Filter (state, predicate, error) = Access.filter predicate error state
+    /// Let the access-controlled function only evaluate once.
     [<CustomOperation("once")>]
     member __.Once (state) = Access.once state
+    /// Let the access-controlled function only evaluate two times.
     [<CustomOperation("twice")>]
     member __.Twice (state) = Access.twice state
+    /// Let the access-controlled function only evaluate a certain amount of times.
     [<CustomOperation("times")>]
     member __.Times (state, count) = Access.times count state
+    /// Make the access-controlled function revokable.
     [<CustomOperation("revokable")>]
     member __.Revokable (state) = Access.revokable state
+    /// Automatically revoke the access-controlled after a specified delay.
     [<CustomOperation("revokedAfter")>]
     member __.RevokedAfter (state, delay) = Access.revokedAfter delay state
+    /// Automatically revoke the access-controlled function when the specified observable emits a value.
     [<CustomOperation("revokedWhen")>]
     member __.RevokedWhen (state, observable) = Access.revokedWhen observable state
+    /// Make the access-controlled function only available during certain dates.
     [<CustomOperation("duringDates")>]
     member __.DuringDates (state, min, max) = Access.duringDates min max state
+    /// Make the access-controlled function only available during certain hours.
     [<CustomOperation("duringHours")>]
     member __.DuringHours (state, min, max) = Access.duringHours min max state
+    /// Adds input-validation to the access-controlled function.
+    [<CustomOperation("satisfy")>]
+    member __.Satisfy (state, spec) = Access.satisfy spec state
 
+/// Exposure values for the `Access<_, _>` type.
 [<AutoOpen>]
 module AccessExposure =
     let access<'a, 'b> = AccessBuilder<'a, 'b> ()
 
+/// Thrown when the access to the result of an access-controlled function is not available.
 exception AccessFailureException of string
 
-type AccessResult<'a> internal (result) =
+/// Result type when the access-controlled function is evaluated.
+type AccessResult<'a> internal (result : Result<'a, string list>) =
+    /// Gets a value indicating whether or not the access-controlled function was evaluated successfully.
     member __.Successful = Result.isOk result 
-    member __.Error =
+    /// Gets the series of errors that occured during the evaluation of the access-controlled function.
+    member __.Errors =
         match result with
-        | Ok _ -> null
-        | Error err -> err
+        | Ok _ -> [||]
+        | Error errs -> Array.ofList errs
+    /// Gets the result of the access-controlled function if the function evaluated successfully, 
+    /// otherwise an `AccessFailureException` is thrown.
     member __.Value = 
       result
       |> Result.getOrElse (fun errs ->
-        let errors = Environment.NewLine + String.Join (Environment.NewLine + " -> ", errs)
-        let message = sprintf "%s: %s" errors
-        raise (ValidationFailureException message))
+        let errors = Environment.NewLine + String.Join (Environment.NewLine + " -> ", Array.ofList errs)
+        let message = sprintf "Access to resource failed: %s" errors
+        raise (AccessFailureException message))
+    /// Tries to get the result of the access-controlled function based on whether the function evaluated successfully.
     member __.TryGetValue (output : outref<'a>) =
       match result with
       | Ok x -> output <- x; true
       | _ -> output <- Unchecked.defaultof<'a>; false
 
+/// Extensions on the `Access<_, _>` type to use in C# context.
 [<Extension>]
 type AccessExtensions () =
+    /// Adds a requirement to the access-controlled function to only allow files with a certain extension.
     [<Extension>]
-    static member Extension (access, fileExtension) =
+    static member FileExtension (access, fileExtension) =
         if fileExtension = null then nullArg "fileExtension"
-        Access.extension fileExtension access
+        Access.fileExtension fileExtension access
+    /// Adds a requirement to the access-controlled function to only allow files with certain extensions.
     [<Extension>]
-    static member Extensions (access, [<ParamArray>] fileExtensions) =
+    static member FileExtensions (access, [<ParamArray>] fileExtensions) =
         if fileExtensions = null then nullArg "fileExtensions"
-        Access.extensions (List.ofArray fileExtensions) access
-
+        Access.fileExtensions (List.ofArray fileExtensions) access
+    /// Let the access-controlled function only evaluate once.
     [<Extension>]
     static member Once (acc) = Access.once acc
-
+    /// Let the access-controlled function only evaluate two times.
     [<Extension>]
     static member Twice (acc) = Access.twice acc
-
+    /// Let the access-controlled function only evaluate a certain amount of times.
     [<Extension>]
     static member Times (acc, count) = Access.times count acc
-
+    /// Make the access-controlled function revokable.
     [<Extension>]
     static member Revokable (acc) = Access.revokable acc
-
+    /// Automatically revoke the access-controlled after a specified delay.
     [<Extension>]
     static member RevokedAfter (acc, delay) = Access.revokedAfter delay acc
-
+    /// Automatically revoke the access-controlled function when the specified observable emits a value.
     [<Extension>]
     static member RevokedWhen (acc, observable) = Access.revokedWhen observable acc
-
+    /// Make the access-controlled function only available during certain dates.
     [<Extension>]
     static member During (acc, min, max) = Access.duringDates min max acc
-
+    /// Make the access-controlled function only available during certain hours.
     [<Extension>]
     static member DuringHours (acc, min, max) = Access.duringHours min max acc
-
+    /// Adds input-validation to the access-controlled function.
+    [<Extension>]
+    static member Satisfy (acc, spec) = Access.satisfy spec acc
+    /// Revokes the access-controlled function.
     [<Extension>]
     static member Revoke (acc : Access<_, _>) = Option.iter (fun f -> f ()) acc.Revokable
-
+    /// Evaluate the access-controlled function.
     [<Extension>]
-    static member Eval (acc : Access<unit, 'b>) =
-      Access.eval () acc |> AccessResult<'b>
-
+    static member Eval (acc : Access<unit, 'b>) = Access.eval () acc |> AccessResult<'b>
+    /// Evaluate the access-controlled function.
     [<Extension>]
-    static member Eval ((acc : Access<'a, 'b>), (input : 'a)) : AccessResult<'b> = 
-        Access.eval input acc |> AccessResult<'b>
-        
+    static member Eval ((acc : Access<'a, 'b>), (input : 'a)) : AccessResult<'b> = Access.eval input acc |> AccessResult<'b>